refactor: update import paths for Task types to improve code organization
